refactor: StdViewerをdomain層からinfra層に移動
- StdViewerをinternal/infra/std.goに移動
- Viewerインターフェースはdomain層に保持
- 「標準出力に出力する」機能は技術的な実装詳細のためinfra層が適切
